+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implement `/api/v1/follow_requests` API endpoint.
 - Implement `/api/v1/follow_requests/{account_id}/authorize` and `/api/v1/follow_requests/{account_id}/reject` API endpoints.
 - Support manual approval of followers.
+- Add `rejected_by` attribute to Mastodon API Relationship object.
 
 ### Changed